requirements.txt
Django==1.8.6
django-tastypie==0.12.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.10.0
wsgiref==0.1.2
